DOC: safetynets: code style pass and cleanup
The current state of the code is not entirely readable, as the roles
aren't clear. The parameters of the functions passed around aren't
completely clear either.

Attempt to clarify the parameters of the functions.

Likewise, structure declarations have been moved to their own header so
as to not clutter the module.